Cleanup global namespace pollution in rl motion runtime. Also add beartype.
