+++ src/bitbots_motion/bitbots_rl_motion/bitbots_rl_motion/__init__.py
@@ -1,0 +1,4 @@
+# Setting up runtime type checking for this package
+from beartype.claw import beartype_this_package
+
+beartype_this_package()
