+++ src/bitbots_motion/bitbots_rl_motion/setup.py
@@ -27,11 +27,11 @@ setup(
     tests_require=["pytest"],
     entry_points={
         "console_scripts": [
-            "walk_node = nodes.walk_node:main",
-            "kick_node = nodes.kick_node:main",
-            "mjlab_walk_node = nodes.mjlab_walk_node:main",
-            "mjlab_getup_node = nodes.mjlab_getup_node:main",
-            "phase_from_transform = nodes.phase_from_transform:main",
+            f"walk_node = {package_name}.nodes.walk_node:main",
+            f"kick_node = {package_name}.nodes.kick_node:main",
+            f"mjlab_walk_node = {package_name}.nodes.mjlab_walk_node:main",
+            f"mjlab_getup_node = {package_name}.nodes.mjlab_getup_node:main",
+            f"phase_from_transform = {package_name}.nodes.phase_from_transform:main",
         ],
     },
 )
